callbacks_working

+++ db/migrate/20181004141902_create_accounts.rb
@@ -1,7 +1,7 @@
 class CreateAccounts < ActiveRecord::Migration[5.2]
   def change
     create_table :accounts do |t|
-      t.decimal :balance
+      t.integer :balance
       t.string :a_type
       t.integer :number
